% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsearch_boinsim.R
\name{gridsearch_boinsim}
\alias{gridsearch_boinsim}
\title{Conduct gridsearch for BOIN dose escalation trials with specified parameters}
\usage{
gridsearch_boinsim(object,trueRate,
                          cohortSize,min_cohortEnrolled,min_cohortEval,maxSubjects,
                          dropout_rate,sf_rate,accrual_rate,
                          window_screen,window_DLT,meanDLTtime,lagPlanned,lagDropout,
                          k.earlystop,n.earlystop,ncores=1)
}
\arguments{
\item{object}{a boinsim object that is output by the function run_boinsim}

\item{trueRate}{vector of true DLT rates for each dose level.}

\item{cohortSize}{number of patients recruited at a time during the dose escalation trial}

\item{min_cohortEnrolled}{minimum number of patients passing the screening at each cohort}

\item{min_cohortEval}{minimum number of patients needed at a cohort if dropouts occur}

\item{maxSubjects}{If this many subjects completed the DLT assessment, the end of escalation is triggered.}

\item{dropout_rate}{probability of dropout}

\item{sf_rate}{probability of screen failure}

\item{accrual_rate}{the rate at which new patients are recruited (per day)}

\item{window_screen}{the screening window (days)}

\item{window_DLT}{the assessment window for DLTs (days)}

\item{meanDLTtime}{mean time for DLT development (among patients who do develop DLT reaction)}

\item{lagPlanned}{the delay or advance (in days) of recruitment for the next cohort from the completion of assessment}

\item{lagDropout}{the delay (in days) of searching for replacement after a dropout is confirmed}

\item{k.earlystop}{can be FALSE or a non-negative integer. When it is a non-negative integer, it controls the maximum number of cohorts enrolled at each dose level. One of k.earlystop and n.earlystop needs to be FALSE.}

\item{n.earlystop}{can be a non-negative integer or FALSE. When not FALSE, it controls the number of subjects enrolled in one dose level which triggers end of escalation.}

\item{ncores}{Number of cores available to run the simulations in parallel.Default to 1, when the jobs will be run in sequence.}
}
\value{
a tibble object with each row storing the boinsim output for a unique parameter combination
}
\description{
Apply \code{run_boinsim} function to all combinations of parameters specified in the argument.
}
\details{
The parameters that take a single value in \code{run_boinsim} could be provided as a single value or a vector containing all the values to be screened.
If more than one set of \code{trueRate} is needed, it could be provided as a data frame with column names: Dose 1, ... Dose k,(index); or a (named) list of vectors.
The "index" column or the names of the list is optional. If available, it will be used as an unique identifier of the true toxicity scenario.

\if{html}{\out{<div class="sourceCode">}}\preformatted{     All unique combinations of the following parameters will be included in the searching grid: trueRate,
     nesting(cohortSize,min_cohortEnrolled,min_cohortEval),maxSubjects,
     dropout_rate,sf_rate,accrual_rate,window_screen,window_DLT,meanDLTtime,lagPlanned,lagDropout,nesting(k.earlystop,n.earlystop). 
     If any of the above parameters is not specified in the argument, the parameter value from the example boinsim \code{object} will be used.
   
     - Note: for the following two groups of parameters, only combinations already present in the argument (parameter values from the same vector index position) will be included:
     (1) \code{cohortSize},\code{min_cohortEnrolled},and\code{min_cohortEval}
     (2) \code{k.earlystop} and \code{n.earlystop}
}\if{html}{\out{</div>}}
}
