% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOC_boinsim.R
\name{plotOC_boinsim}
\alias{plotOC_boinsim}
\title{Generate operating characteristics plots.}
\usage{
plotOC_boinsim(object,type,which=c(1,2,5,6),
              cohortSize,min_cohortEnrolled,min_cohortEval,maxSubjects,
              dropout_rate,sf_rate,accrual_rate,
              window_screen,window_DLT,meanDLTtime,lagPlanned,lagDropout,
              k.earlystop,n.earlystop)
}
\arguments{
\item{object}{an object output by run_boinsim or gridsearch_boinsim.}

\item{type}{'dose' or 'interval'. Whether to plot by groups of doses or intervals (below, within, or above the target interval).}

\item{which}{a subset of the numbers 1:6, by default 1,2,5,6, referring to:
1. true toxicity rate at each dose level (a line plot if only one toxicity scenario is include, a jitterplot otherwise);
2. # of DLTs
3. # of subjects recruited
4. # of subjects enrolled
5. # of subjects evaluable
6. MTD selection percentage}

\item{cohortSize}{number of patients recruited at a time during the dose escalation trial}

\item{min_cohortEnrolled}{minimum number of patients passing the screening at each cohort}

\item{min_cohortEval}{minimum number of patients needed at a cohort if dropouts occur}

\item{maxSubjects}{If this many subjects completed the DLT assessment, the end of escalation is triggered.}

\item{dropout_rate}{probability of dropout}

\item{sf_rate}{probability of screen failure}

\item{accrual_rate}{the rate at which new patients are recruited (per day)}

\item{window_screen}{the screening window (days)}

\item{window_DLT}{the assessment window for DLTs (days)}

\item{meanDLTtime}{mean time for DLT development (among patients who do develop DLT reaction)}

\item{lagPlanned}{the delay or advance (in days) of recruitment for the next cohort from the completion of assessment}

\item{lagDropout}{the delay (in days) of searching for replacement after a dropout is confirmed}

\item{k.earlystop}{can be FALSE or a non-negative integer. When it is a non-negative integer, it controls the maximum number of cohorts enrolled at each dose level. One of k.earlystop and n.earlystop needs to be FALSE.}

\item{n.earlystop}{can be a non-negative integer or FALSE. When not FALSE, it controls the number of subjects enrolled in one dose level which triggers end of escalation.}
}
\description{
Generate plots for true toxicity rates, dose-wise or interval-wise number of DLTs, number of recruited/enrolled/evaluable subjects, and MTD selection percentage.
}
\details{
\itemize{
\item If \code{object} is an output by run_boinsim, \code{rawOutput} has to be set to TRUE in run_boinsim. The plots will be a summary of all simulation trials.
There is no need to specify parameters other than \code{type} and \code{which}.
- If \code{object} is an output by gridsearch_boinsim, the plots will be a summary across all toxicity scenarios, but only for one combination of other parameters as specified by the rest of the arguments.
For the arguments that only take one value in \code{object} (not subject to gridsearch), there is no need to specify here again.
}
}
