% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_boinsim.R
\name{run_boinsim}
\alias{run_boinsim}
\title{Simulate BOIN dose escalation with patient dropout and timeline monitoring}
\usage{
run_boinsim(targetRate, trueRate, cohortSize, min_cohortEnrolled,
       min_cohortEval,maxSubjects,dropout_rate,sf_rate=0.2,accrual_rate=0.1,
       window_screen=14,window_DLT=28,meanDLTtime=7,
       lagPlanned=0,lagDropout=0,k.earlystop=3,n.earlystop=FALSE,
       p.saf=0.6*targetRate,p.tox=1.227*targetRate,cutoff.eli=0.95,
       ntrial=100,seed=2022,rawOutput=FALSE)
}
\arguments{
\item{targetRate}{The target rate for the BOIN design}

\item{trueRate}{vector of true DLT rates for each dose level}

\item{cohortSize}{number of patients recruited at a time during the dose escalation trial}

\item{min_cohortEnrolled}{minimum number of patients passing the screening at each cohort}

\item{min_cohortEval}{minimum number of patients needed at a cohort if dropouts occur}

\item{maxSubjects}{If this many subjects completed the DLT assessment, the end of escalation is triggered.}

\item{dropout_rate}{probability of dropout}

\item{sf_rate}{probability of screen failure}

\item{accrual_rate}{the rate at which new patients are recruited (per day)}

\item{window_screen}{the screening window (days)}

\item{window_DLT}{the assessment window for DLTs (days)}

\item{meanDLTtime}{mean time for DLT development (among patients who do develop DLT reaction)}

\item{lagPlanned}{the delay or advance (in days) of recruitment for the next cohort from the completion of assessment}

\item{lagDropout}{the delay (in days) of searching for replacement after a dropout is confirmed}

\item{k.earlystop}{can be FALSE or a non-negative integer. When it is a non-negative integer, it controls the maximum number of cohorts enrolled at each dose level. One of k.earlystop and n.earlystop needs to be FALSE.}

\item{n.earlystop}{can be a non-negative integer or FALSE. When not FALSE, it controls the number of subjects enrolled in one dose level which triggers end of escalation.}

\item{p.saf}{the highest toxicity rate that is deemed subtherapeutic (lower boundary of the target interval)}

\item{p.tox}{the lowest toxicity rate that is deemed overly toxic (upper boundary of the target interval)}

\item{cutoff.eli}{The elimination threshold used to generate the BOIN decision table. Default to 0.95}

\item{ntrial}{number of simulation trials to run. Default to 100.}

\item{seed}{the random seed for simulation. Defaul to 2022.}

\item{rawOutput}{whether to return the raw output. Default to FALSE, when only a summary of all simulation trials is returned.}
}
\value{
If rawOutput==TRUE, the output for a single simulation trial is returned in a list, including:
(1) DLTS: number of patients experiencing DLT at each dose level
(2) DLTs_interval: number of patients experiencing DLT at each interval level (below, within, or above the target interval)
(3) Subjects/Subjects_perc: number/percentage of patients recruited at each dose level
(4) SubjectsEnrolled/SubjectsEnrolled_perc: number/percentage of patients who pass the screening at each dose level
(5) SubjectsCompleted/SubjectsCompleted_perc: number/percentage of evaluable patients at each dose level
(6) SubjectsInterval/SubjectsInterval_perc: number/percentage of patients recruited at each interval level
(7) SubjectsEnrolledInterval/SubjectsEnrolledInterval_perc: number/percentage of patients who pass the screening at each interval level
(8) SubjectsCompletedInterval/SubjectsCompletedInterval_perc: number/percentage of evaluable patients at each interval level
(9) Cohorts: number of cohorts assigned at each dose level
(10) CohortsInterval: number of cohorts assigned at each interval level
(11) avg_DurationCohort: average duration for a cohort to complete, from searching patients to completion of assessment
(12) DurationTrial: total duration of the trial
(13) DosePath: the path of dose escalation/deescaltion/stay during the trial
(14) DoseSelected: dose selected as MTD
(15) DoseSelected_interval: whether the selected MTD is below, within, or above the target interval

\if{html}{\out{<div class="sourceCode">}}\preformatted{    If rawOutput==FALSE (default), a mean over all simulation trials are calculated and output for item 1~12, in addition the following items are included in the output list:
    (13) SlcPerc_dose: MTD selection percentage for 'All excluded' and all dose levels
    (14) SlcPerc_interval: MTD selection percentage for 'All excluded' and below, within, or above the target interval
    
    A dataframe of simulation setup (sim.setup) will be included in the output list for both versions.
}\if{html}{\out{</div>}}
}
\description{
Simulate BOIN dose escalation with patient dropout and timeline monitoring
}
\author{
Lu, Yixing; Henner, William
}
