% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueToxRate.R
\name{trueToxRate}
\alias{trueToxRate}
\title{Simulate ground truth toxicity scenarios for BOIN dose escalation trials simulation}
\usage{
trueToxRate(targetRate,ndose,sigma0,mu,sigma1,nsim,seed=NULL)
}
\arguments{
\item{targetRate}{The target toxicity rate used in the BOIN design}

\item{ndose}{number of dose levels}

\item{sigma0}{standard deviation of a normal distribution N(qnorm(\code{targetRate}),\code{sigma0}) from which the true toxicity rate of MTD is sampled. See details.}

\item{mu}{mean of the normal distribution N(\code{mu},\code{sigma1}^2) from which the toxicity rate difference between adjacent dose levels is sampled. See details.}

\item{sigma1}{standard deviation of the normal distribution N(\code{mu},\code{sigma1}^2). See details.}

\item{nsim}{number of scenarios to simulate}

\item{seed}{the random seed for simulation}
}
\value{
\code{trueToxRate()} returns a list object, including
(1) \code{$trueRate}: a matrix of size (\code{nsim},\code{ndose}) with the true toxicity rate for each dose level for all the simulations.
(2) \code{$MTD}: a vector of length \code{nsim} storing the MTD level of each simulated dose-toxicity curve. 0 if all doses are toxic, and \code{ndose+1} if MTD is not reached.
(3)\code{$gap_around_target}: a vector of length \code{nsim} storing the mean toxicity rate difference around the MTD for each simulated dose-toxicity curve.
}
\description{
Use this function to simulate toxicity-dose curves.
}
\details{
\itemize{
\item True MTD is randomly sampled with equal probability from all dose levels, or 'all doses toxic', or 'MTD not reached'.
- The true toxicity rate of the randomly selected MTD is allowed to deviate from the target rate. This is achieved by
setting p(j) = pnorm(err_j), where err_j is drawn from the normal distribution N(qnorm(\code{targetRate}),\code{sigma0}).
\code{sigma0} controls the variation of the true toxicty rate of MTD from the target rate.
- The difference in toxicity rate between adjacent dose levels (steepness of dose-toxicity curves) is controlled by \code{mu} and \code{sigma1}.
p(j-1)=pnorm(qnorm(p(j))-err_{j-1}^2), p(j+1)=pnorm(qnorm(p(j))+err_{j+1}^2), where err_{j-1} and err_{j+1}
are drawn from the normal distribution N(\code{mu},\code{sigma1}^2).
Additional constraint is applied to ensure p(j) is closest to the target rate.
}
}
\references{
Liu S. and Yuan, Y. (2015). Bayesian Optimal Interval Designs for Phase I
Clinical Trials, \emph{Journal of the Royal Statistical Society: Series C}, 64, 507-523.
}
\author{
Lu, Yixing
}
